+++ GRD/GRD/Controllers/UsersController.cs
@@ -23,6 +23,10 @@ namespace GRD.Controllers
         // GET: Users
         public async Task<IActionResult> Index(string searchString, string filterType)
         {
+            if (!IsAuthorized())
+            {
+                return Unauthorized();
+            }
             ViewData["CurrentFilter"] = searchString;
             ViewData["FilterType"] = filterType;
             var isAdmin = HttpContext.Session.GetString("isAdmin") == "true" ? true : false;
@@ -49,27 +53,13 @@ namespace GRD.Controllers
             return View(await users.AsNoTracking().ToListAsync());
         }
 
-        // GET: Users/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var user = await _context.Users
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
-
-            return View(user);
-        }
-
         // GET: Users/Create
         public IActionResult Create()
         {
+            if (!IsAuthorized())
+            {
+                return Unauthorized();
+            }
             return View();
         }
 
@@ -80,6 +70,10 @@ namespace GRD.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Username,Password,Address,Id,Gender,IsAdmin")] User user)
         {
+            if (!IsAuthorized())
+            {
+                return Unauthorized();
+            }
             if (ModelState.IsValid)
             {
                 _context.Add(user);
@@ -92,6 +86,10 @@ namespace GRD.Controllers
         // GET: Users/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!IsAuthorized())
+            {
+                return Unauthorized();
+            }
             if (id == null)
             {
                 return NotFound();
@@ -112,6 +110,10 @@ namespace GRD.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Username,Password,Address,Id,Gender,IsAdmin")] User user)
         {
+            if (!IsAuthorized())
+            {
+                return Unauthorized();
+            }
             if (id != user.Id)
             {
                 return NotFound();
@@ -143,6 +145,10 @@ namespace GRD.Controllers
         // GET: Users/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!IsAuthorized())
+            {
+                return Unauthorized();
+            }
             if (id == null)
             {
                 return NotFound();
@@ -173,6 +179,7 @@ namespace GRD.Controllers
             }
             HttpContext.Session.SetString("isAdmin", user.IsAdmin ? "true" : "false");
             HttpContext.Session.SetString("username", user.Username);
+            HttpContext.Session.SetString("isLogin", "true");
 
             if (user.IsAdmin)
             {
@@ -181,11 +188,25 @@ namespace GRD.Controllers
             return RedirectToAction("Index", "Home", null);
         }
 
+        [HttpGet]
+        public IActionResult Logout()
+        {
+            HttpContext.Session.SetString("isAdmin", "false");
+            HttpContext.Session.SetString("username", "");
+            HttpContext.Session.SetString("isLogin", "false");
+
+            return RedirectToAction("Index", "Home", null);
+        }
+
         // POST: Users/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!IsAuthorized())
+            {
+                return Unauthorized();
+            }
             var user = await _context.Users.FindAsync(id);
             _context.Users.Remove(user);
             await _context.SaveChangesAsync();
@@ -195,6 +216,12 @@ namespace GRD.Controllers
         private bool UserExists(int id)
         {
             return _context.Users.Any(e => e.Id == id);
+        }
+
+        private bool IsAuthorized()
+        {
+            return (HttpContext.Session.GetString("isAdmin") == "true" ? true : false) &&
+                (HttpContext.Session.GetString("isLogin") == "true" ? true : false);
         }
     }
 }
